+++ src/main/kotlin/com/coder/toolbox/store/StoreKeys.kt
@@ -10,7 +10,10 @@ internal const val APP_NAME_AS_TITLE = "useAppNameAsTitle"
 
 internal const val BINARY_SOURCE = "binarySource"
 
+@Deprecated("Use BINARY_DESTINATION instead", replaceWith = ReplaceWith("BINARY_DESTINATION"))
 internal const val BINARY_DIRECTORY = "binaryDirectory"
+
+internal const val BINARY_DESTINATION = "binaryDestination"
 
 internal const val DISABLE_SIGNATURE_VALIDATION = "disableSignatureValidation"
 
@@ -18,13 +21,9 @@ internal const val FALLBACK_ON_CODER_FOR_SIGNATURES = "signatureFallbackStrategy
 
 internal const val HTTP_CLIENT_LOG_LEVEL = "httpClientLogLevel"
 
-internal const val BINARY_NAME = "binaryName"
-
 internal const val DATA_DIRECTORY = "dataDirectory"
 
 internal const val ENABLE_DOWNLOADS = "enableDownloads"
-
-internal const val ENABLE_BINARY_DIR_FALLBACK = "enableBinaryDirectoryFallback"
 
 internal const val HEADER_COMMAND = "headerCommand"